WebSockets/tests/submissions/Opera/cookies/001.html

<!doctype html>
<title>WebSockets: Cookie in request</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  if (window.WebSocket) {
    document.cookie = 'ws_test=test; Path=/';
  }
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo-cookie');
  ws.onopen = debug;
  ws.onmessage = t.step_func(function(e) {
    assert_regexp(e.data, /ws_test=test/);
    ws.close();
    t.done();
  });
  ws.onerror = ws.onclose = t.step_func(function() {assert_unreached()});
});
// remove cookie
document.cookie = 'ws_test=; Path=/; Expires=Sun, 06 Nov 1994 08:49:37 GMT';
</script>
